+++ LovedLabels/Framework/LoveLabelConfig.cs
@@ -1,0 +1,15 @@
+namespace LovedLabels.Framework
+{
+    public class LoveLabelConfig
+    {
+        public string AlreadyPettedLabel { get; set; } = "Is Loved";
+
+        public string NeedsToBePettedLabel { get; set; } = "Needs Love";
+
+        /*public LoveLabelConfig()
+        {
+            AlreadyPettedLabel = "Is Loved";
+            NeedsToBePettedLabel = "Needs Love";
+        }*/
+    }
+}
