+++ LovedLabels/LovedLabels.cs
@@ -1,14 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using LovedLabels.Framework;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Content;
+using Microsoft.Xna.Framework.Graphics;
+using StardewModdingAPI;
+using StardewModdingAPI.Events;
 using StardewValley;
 using StardewValley.Characters;
 using StardewValley.Menus;
-using StardewModdingAPI;
-using StardewModdingAPI.Events;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Content;
 
 namespace LovedLabels
 {
@@ -65,20 +66,20 @@ namespace LovedLabels
             {
                 // Following values could use tweaking, no idea wtf is going on here
                 RectangleF animalBoundaries = new RectangleF(animal.position.X, animal.position.Y - animal.sprite.getHeight(), animal.sprite.getWidth() * 3 + animal.sprite.getWidth() / 1.5f, animal.sprite.getHeight() * 4);
-                
+
                 if (animalBoundaries.Contains(mousePos.X * Game1.tileSize, mousePos.Y * Game1.tileSize))
                 {
                     hoverPetLabel = animal.wasPet ? config.AlreadyPettedLabel : config.NeedsToBePettedLabel;
                 }
             }
-            
+
             foreach (NPC npc in currentLocation.characters)
             {
                 if (npc is Pet)
                 {
                     // Following values could use tweaking, no idea wtf is going on here
                     RectangleF petBoundaries = new RectangleF(npc.position.X, npc.position.Y - npc.sprite.getHeight() * 2, npc.sprite.getWidth() * 3 + npc.sprite.getWidth() / 1.5f, npc.sprite.getHeight() * 4);
-                    
+
                     if (petBoundaries.Contains(mousePos.X * Game1.tileSize, mousePos.Y * Game1.tileSize))
                     {
                         //bool wasPet = (bool)typeof(Pet).GetField("wasPetToday", BindingFlags.Instance | BindingFlags.NonPublic).GetValue(npc as Pet);
@@ -127,41 +128,6 @@ namespace LovedLabels
             int sourceY = (hoverText == config.AlreadyPettedLabel) ? 0 : 32;
             Vector2 heartpos = new Vector2(x + textSize.X + halfHeartSize, y + halfHeartSize);
             b.Draw(hearts, heartpos, new Rectangle(0, sourceY, 32, 32), Color.White);
-        }    
-    }
-
-    public class LoveLabelConfig
-    {
-        public string AlreadyPettedLabel { get; set; } = "Is Loved";
-
-        public string NeedsToBePettedLabel { get; set; } = "Needs Love";
-
-        /*public LoveLabelConfig()
-        {
-            AlreadyPettedLabel = "Is Loved";
-            NeedsToBePettedLabel = "Needs Love";
-        }*/
-    }
-
-    public class RectangleF
-    {
-        public float X;
-        public float Y;
-        public float Width;
-        public float Height;
-
-        public RectangleF(float x, float y, float width, float height)
-        {
-            X = x;
-            Y = y;
-            Width = width;
-            Height = height;
-        }
-
-        public bool Contains(float xPos, float yPos)
-        {
-            return this.X <= xPos && xPos < this.X + this.Width && this.Y <= yPos && yPos < this.Y + this.Height;
         }
     }
-
 }
