+++ LovedLabels/LovedLabels.csproj
@@ -36,6 +36,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="LovedLabels.cs" />
+    <Compile Include="Framework\LoveLabelConfig.cs" />
+    <Compile Include="Framework\RectangleF.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
